accountページのリファクタリング
