+++ src/app/account/page.tsx
@@ -1,25 +1,13 @@
 "use client";
-import { auth } from "@/app/firebase";
-import NameUpdate from "@/Components/NameUpdate/NameUpdate";
-import Notification from "@/Components/Notification/Notification";
-import { signInAsGuest } from "@/utils/Auth/signInAnonymously";
-import { signInWithMail } from "@/utils/Auth/signInWithMail";
-import { signUpWithGoogleAccount } from "@/utils/Auth/signUpWithGoogleAccount";
-import { signUpWithMail } from "@/utils/Auth/signUpWithMail";
+import AuthForm from "@/Components/Account/AuthForm";
+import LoggedInView from "@/Components/Account/LoggedInView";
 import { useUser } from "@/utils/UserContext";
 import Box from "@mui/material/Box";
-import Button from "@mui/material/Button";
 import MuiCard from "@mui/material/Card";
 import CssBaseline from "@mui/material/CssBaseline";
-import Divider from "@mui/material/Divider";
 import Stack from "@mui/material/Stack";
 import { styled } from "@mui/material/styles";
-import TextField from "@mui/material/TextField";
-import ToggleButton from "@mui/material/ToggleButton";
-import ToggleButtonGroup from "@mui/material/ToggleButtonGroup";
 import Typography from "@mui/material/Typography";
-import { signOut } from "firebase/auth";
-import React, { useState } from "react";
 
 const Card = styled(MuiCard)(({ theme }) => ({
   display: "flex",
@@ -37,53 +25,8 @@ const Card = styled(MuiCard)(({ theme }) => ({
   },
 }));
 
-const CenteredToggleButtonGroup = styled(ToggleButtonGroup)({
-  display: "flex",
-  justifyContent: "center",
-  marginBottom: "16px",
-});
-
-const RoundedButton = styled(Button)(({ theme }) => ({
-  borderRadius: "50px",
-  padding: theme.spacing(1.5, 4),
-}));
-
 export default function Account() {
-  const [name, setName] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [formMode, setFormMode] = useState<"register" | "login">("register");
-  const [open, setOpen] = useState(false);
-
   const { user } = useUser();
-
-  const handleRegisterSubmit = async (event: React.FormEvent) => {
-    event.preventDefault();
-    await signUpWithMail(email, password, name);
-  };
-
-  const handleLoginSubmit = async (event: React.FormEvent) => {
-    event.preventDefault();
-    await signInWithMail(email, password);
-  };
-
-  const handleGoogleLogin = async () => {
-    await signUpWithGoogleAccount();
-  };
-
-  const handleGuestLogin = async () => {
-    await signInAsGuest();
-  };
-
-  const handleLogout = async () => {
-    try {
-      await signOut(auth);
-      console.log("Signed out");
-    } catch (error) {
-      console.error("errorCode:", (error as Error)?.name);
-      console.error("errorMessage:", (error as Error)?.message);
-    }
-  };
 
   return (
     <>
@@ -98,161 +41,13 @@ export default function Account() {
           <Typography variant="h4" textAlign="center">
             TODO-REAL
           </Typography>
-          {!user && (
-            <CenteredToggleButtonGroup
-              value={formMode}
-              exclusive
-              onChange={(event, newValue) => {
-                if (newValue) {
-                  setFormMode(newValue);
-                }
-              }}
-              aria-label="form mode"
-            >
-              <ToggleButton value="register" aria-label="register">
-                新規登録
-              </ToggleButton>
-              <ToggleButton value="login" aria-label="login">
-                ログイン
-              </ToggleButton>
-            </CenteredToggleButtonGroup>
-          )}
           <Box sx={{ display: "flex", flexDirection: "column", gap: 2 }}>
             {user ? (
-              <>
-                <Typography>ログイン中: {user.name}</Typography>
-
-                {/* ゲストの場合は通知機能を利用しない */}
-                {user.loginType !== "Guest" && <Notification />}
-
-                {/* メール認証が完了していない場合に表示 */}
-                {!user.isMailVerified && (
-                  <Typography color="error">
-                    メールに届いた認証リンクを確認してください。
-                    <br />
-                    認証が完了するまで閲覧以外の機能は制限されます。
-                  </Typography>
-                )}
-
-                <RoundedButton variant="contained" onClick={handleLogout}>
-                  ログアウト
-                </RoundedButton>
-
-                {/* ゲストの場合は名前を変更しない */}
-                {user.loginType !== "Guest" && (
-                  <>
-                    <RoundedButton
-                      variant="contained"
-                      onClick={() => setOpen(true)}
-                    >
-                      名前を変更
-                    </RoundedButton>
-                    <NameUpdate open={open} setOpen={setOpen} />
-                  </>
-                )}
-              </>
+              // ログイン中の場合
+              <LoggedInView />
             ) : (
-              <>
-                {formMode === "register" ? (
-                  <>
-                    <Typography>新規登録</Typography>
-                    <form onSubmit={handleRegisterSubmit}>
-                      <Box
-                        sx={{
-                          display: "flex",
-                          flexDirection: "column",
-                          gap: 2,
-                        }}
-                      >
-                        <TextField
-                          label="Username"
-                          value={name}
-                          onChange={(e) => setName(e.target.value)}
-                          fullWidth
-                          required
-                        />
-                        <TextField
-                          label="Email"
-                          type="email"
-                          value={email}
-                          onChange={(e) => setEmail(e.target.value)}
-                          fullWidth
-                          required
-                        />
-                        <TextField
-                          label="Password"
-                          type="new-password"
-                          value={password}
-                          onChange={(e) => setPassword(e.target.value)}
-                          fullWidth
-                          required
-                        />
-                        <RoundedButton
-                          type="submit"
-                          variant="contained"
-                          fullWidth
-                        >
-                          アカウント作成
-                        </RoundedButton>
-                      </Box>
-                    </form>
-                  </>
-                ) : (
-                  <>
-                    <Typography>ログイン</Typography>
-                    <form onSubmit={handleLoginSubmit}>
-                      <Box
-                        sx={{
-                          display: "flex",
-                          flexDirection: "column",
-                          gap: 2,
-                        }}
-                      >
-                        <TextField
-                          label="Email"
-                          type="email"
-                          value={email}
-                          onChange={(e) => setEmail(e.target.value)}
-                          fullWidth
-                          required
-                        />
-                        <TextField
-                          label="Password"
-                          type="current-password"
-                          value={password}
-                          onChange={(e) => setPassword(e.target.value)}
-                          fullWidth
-                          required
-                        />
-                        <RoundedButton
-                          type="submit"
-                          variant="contained"
-                          fullWidth
-                        >
-                          ログイン
-                        </RoundedButton>
-                      </Box>
-                    </form>
-                  </>
-                )}
-                <>
-                  <Divider>または</Divider>
-                  <RoundedButton
-                    fullWidth
-                    variant="outlined"
-                    onClick={handleGoogleLogin}
-                  >
-                    Googleでログイン
-                  </RoundedButton>
-                  <RoundedButton
-                    fullWidth
-                    variant="outlined"
-                    onClick={handleGuestLogin}
-                  >
-                    ゲストログイン
-                  </RoundedButton>
-                </>
-              </>
+              // ログインしていない場合
+              <AuthForm />
             )}
           </Box>
         </Card>
